python-backend/requirements.txt
openai-agents
pydantic
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain_core
langchain_community
langchain_openai
langchain_text_splitters
streamlit
altair
pandas
pypdf